Further fixes for #104, #105, #109, and #117
Added the "SendToEvernoteActivity" directory to the header search
paths. Including it with the directory worked for CocoaPods with
Objective-C, but not for Swift.

This may cause an error for those including it via a framework, but
since CocoaPods can't deal with this consistently, flattening the
heirarchy will be a better long term fix

Also, bumped to 2.0.6